Add a template-functions.php file
